+++ src/CSharpLanguageServer/Types.fs
@@ -12,13 +12,21 @@ type CSharpDebugConfiguration =
         { solutionLoadDelay = None
           debugMode = None }
 
+type CSharpCompletionConfiguration =
+    { showItemsFromUnimportedNamespaces: bool option
+      showNameSuggestions: bool option }
+
+    static member Default =
+        { showItemsFromUnimportedNamespaces = None
+          showNameSuggestions = None }
+
 type CSharpConfiguration =
     { logLevel: string option
       applyFormattingOptions: bool option
       analyzersEnabled: bool option
       useMetadataUris: bool option
       razorSupport: bool option
-      completionShowItemsFromUnimportedNamespaces: bool option
+      completion: CSharpCompletionConfiguration option
       locale: string option
       debug: CSharpDebugConfiguration option
       solutionPathOverride: string option }
@@ -35,7 +43,7 @@ type CSharpConfiguration =
           analyzersEnabled = None
           useMetadataUris = None
           razorSupport = None
-          completionShowItemsFromUnimportedNamespaces = None
+          completion = None
           locale = None
           debug = None
           solutionPathOverride = None }
@@ -48,9 +56,17 @@ let mergeCSharpConfiguration (oldConfig: CSharpConfiguration) (newConfig: CSharp
       analyzersEnabled = newConfig.analyzersEnabled |> Option.orElse oldConfig.analyzersEnabled
       useMetadataUris = newConfig.useMetadataUris |> Option.orElse oldConfig.useMetadataUris
       razorSupport = newConfig.razorSupport |> Option.orElse oldConfig.razorSupport
-      completionShowItemsFromUnimportedNamespaces =
-        newConfig.completionShowItemsFromUnimportedNamespaces
-        |> Option.orElse oldConfig.completionShowItemsFromUnimportedNamespaces
+      completion =
+        match newConfig.completion with
+        | Some newCompletion ->
+            Some
+                { showItemsFromUnimportedNamespaces =
+                    newCompletion.showItemsFromUnimportedNamespaces
+                    |> Option.orElse (oldConfig.completion |> Option.bind _.showItemsFromUnimportedNamespaces)
+                  showNameSuggestions =
+                    newCompletion.showNameSuggestions
+                    |> Option.orElse (oldConfig.completion |> Option.bind _.showNameSuggestions) }
+        | None -> oldConfig.completion
       locale = newConfig.locale |> Option.orElse oldConfig.locale
       debug =
         match newConfig.debug with